Map ecdsa/ed25519 args to create secret
